CrewAI-imobiliaria/requirements.txt
crewai
crewai-tools
groq
langchain
langchain_community
langchain_groq
python-dotenv
duckduckgo_search